Move third-party sources to a new ext/ directory
