+++ ext/ksw2_extz2_sse.c
