+++ ext/ssw.c
