+++ ext/ssw_cpp.h
