+++ ext/xxhash.c
