+++ ext/xxhash.h
